this is actually pointless
The server application is mostly synchronous. `cycle-workers!` blocks the main loop until it returns, so no matter how long we sleep here, we won't register any additional workers.

Punting on this for now. There is at least code in the worker processes where they check for process suspension and exit, so between that and killing the busy workers, I think we're covered.