bugfix: move process suspension checking up in priority
What's happening here is that when a worker comes back from process suspension, the first thing it checks is whether it hasn't received a heartbeat from the server, then it jumps right back into exchanging heartbeats and working.

Instead, the worker should check for process suspension FIRST, and exit if that's the case.