increase the time a disconnected worker will hang around from 10 to 30 seconds
I've noticed some issues where it takes the server a little over 10 seconds
before it sends the first heartbeat, so the workers give up and commit seppuku
when all they had to do was just wait a little bit longer.

I imagine we'll still have to tweak this in the future, but I'm pretty sure 30
is a better number than 10, at least.